requirements.txt
Django
ipython
Pillow
psycopg2-binary
python-dotenv
django-filter
djangorestframework
